Move TimeTableEntry type to shared types
